refactor: animation 폴더 따로 생성하여 기존 파일 이동
